sentiment_analysis_albert_emoji/requirements.txt
tensorflow1.15
sentencepiece
